Fix `eager_fire_and_forget` (#14)
The coroutine state may not be completely constructed by the time the task constructor is called, which can lead to weird behavior if the constructor calls 'resume()`. Instead it is better to use `await_ready()` in this case.

The standard in [§9.5.4#7](https://timsong-cpp.github.io/cppwp/dcl.fct.def.coroutine#7) says "The call to get_return_object is sequenced before the call to initial_suspend and is invoked at most once.", but I couldn't find anything if this also applies to the construction of the return object.